+++ acme.py
@@ -21,8 +21,9 @@ description = 'ACME ' + version + ' - An open source CSE Middleware for Educatio
 # Handle command line arguments
 def parseArgs():
 	parser = argparse.ArgumentParser(description=description)
-	parser.add_argument('--config', action='store', dest='configfile', default=defaultConfigFile, help='specify configuration file')
+	parser.add_argument('--config', action='store', dest='configfile', default=defaultConfigFile, help='specify the configuration file')
 	parser.add_argument('--db-reset', action='store_true', dest='dbreset', default=None, help='reset the DB when starting the CSE')
+	parser.add_argument('--db-storage', action='store', dest='dbstoragemode', default=None, choices=[ 'memory', 'disk' ], type=str.lower, help='specify the DB´s storage mode')
 	parser.add_argument('--log-level', action='store', dest='loglevel', default=None, choices=[ 'info', 'error', 'warn', 'debug', 'off'], type=str.lower, help='set the log level, or turn logging off')
 	parser.add_argument('--import-directory', action='store', dest='importdirectory', default=None, help='specify the import directory')
 	return parser.parse_args()